Use test to toggle register/login
