--- client/src/components/User.js
@@ -1,90 +1,0 @@
-import React from 'react';
-import Card from '@material-ui/core/Card';
-import CardContent from '@material-ui/core/CardContent';
-import useMediaQuery from '@material-ui/core/useMediaQuery'
-import Container from '@material-ui/core/Container';
-import IconButton from '@material-ui/core/IconButton';
-import ArrowBackIcon from '@material-ui/icons/ArrowBack';
-import { Link } from 'react-router-dom';
-import TextField from '@material-ui/core/TextField';
-import Grid from '@material-ui/core/Grid';
-import Button from '@material-ui/core/Button';
-import Typography from '@material-ui/core/Typography';
-
-
-export default function User(props) {
-    let entryWidth = {   
-        width:'95vw',
-    }
-
-    const cardStyle = {
-        backgroundColor: '#d1c4e9',
-        height: "80vh",
-        marginBottom: "calc(20vh-5em)"
-    }
-    const textfieldStyle = {
-        width: "80%",
-        marginTop: "1em"
-    }
-
-    if(useMediaQuery('(min-width:37em)')){
-        entryWidth = {   
-            width:'35em',
-        }
-    }
-
-  return (
-    <Container style={entryWidth}>
-        <Card style={cardStyle}>
-        <CardContent>
-            <Grid
-                container
-                direction="row"
-                justify="space-between"
-                alignItems="center"
-            >   
-            <Link to="/">
-                <IconButton color="secondary" size="small">
-                    <ArrowBackIcon />
-                </IconButton>
-            </Link>
-            <Typography variant="h5" component="h2">
-                User
-            </Typography>
-            <div>
-                
-            </div>
-            </Grid>
-            <form>
-                <TextField
-                        required
-                        id="outlined-required"
-                        label="User"
-                        defaultValue=""
-                        variant="outlined"
-                        fullWidth
-                        color="secondary"
-                        style={textfieldStyle}
-                />
-                <TextField
-                        required
-                        id="outlined-required"
-                        label="Password"
-                        defaultValue=""
-                        variant="outlined"
-                        fullWidth
-                        color="secondary"
-                        style={textfieldStyle}
-                        type="password"
-                />
-                <Button variant="outlined" color="secondary" style={textfieldStyle}>
-                    Log in
-                </Button>
-
-                </form>
-
-            </CardContent>
-        </Card>
-    </Container>
-  );
-}
