+++ client/src/components/entry/EntryCard.js
