+++ client/src/components/entry/EntryCardContent.js
