+++ client/src/components/entry/EntryCardToggle.js
